[refactor] likereview 테스트 수정
