+++ src/test/java/com/inq/wishhair/wesharewishhair/review/domain/likereview/LikeReviewTest.java
@@ -28,50 +28,12 @@ public class LikeReviewTest {
         Review review = ReviewFixture.A.toEntity(user, hairStyle);
 
         //when
-        LikeReview likeReview = LikeReview.addLike(user, review);
+        LikeReview likeReview = LikeReview.addLike(1L, 1L);
 
         //then
         assertAll(
-                () -> assertThat(likeReview.getReview()).isEqualTo(review),
-                () -> assertThat(likeReview.getUser()).isEqualTo(user)
+                () -> assertThat(likeReview.getReviewId()).isEqualTo(1L),
+                () -> assertThat(likeReview.getUserId()).isEqualTo(1L)
         );
-    }
-
-    @Nested
-    @DisplayName("현재 좋아요의 유저와 입력된 유저가 같은지 확인한다")
-    class isSameUser {
-
-        private LikeReview likeReview;
-
-        @BeforeEach
-        void setUp() {
-            //given
-            User user = Mockito.mock(User.class);
-            HairStyle hairStyle = HairStyleFixture.A.toEntity();
-            Review review = ReviewFixture.A.toEntity(user, hairStyle);
-            likeReview = LikeReview.addLike(user, review);
-
-            given(user.getId()).willReturn(1L);
-        }
-
-        @Test
-        @DisplayName("같은 유저를 입력받아 true 를 반환한다")
-        void test2() {
-            //when
-            boolean result = likeReview.isSameUser(1L);
-
-            //then
-            assertThat(result).isTrue();
-        }
-
-        @Test
-        @DisplayName("다른 유저를 입력받아 false 를 반환한다")
-        void test3() {
-            //when
-            boolean result = likeReview.isSameUser(2L);
-
-            //then
-            assertThat(result).isFalse();
-        }
     }
 }
